+++ app/src/main/java/com/pictureair/photopass/widget/RegisterOrForgetCallback.java
@@ -9,6 +9,6 @@ public interface RegisterOrForgetCallback {
     void onFai(int StringId);//操作失败
     void onFai(String StringId);//操作失败
     void onSuccess();//操作成功
-    void countDown(int i);//倒计时
+    void countDown(long i);//倒计时
     void nextPageForget();
 }
